make it a DUU type smoke test
