+++ test/rekt/features/pingsource/readyness.go
@@ -19,25 +19,12 @@ package pingsource
 import (
 	"knative.dev/eventing/test/rekt/features"
 	"knative.dev/eventing/test/rekt/resources/pingsource"
-	"knative.dev/eventing/test/rekt/resources/svc"
-	duckv1 "knative.dev/pkg/apis/duck/v1"
 	"knative.dev/reconciler-test/pkg/feature"
 )
 
-// PingSourceGoesReady returns a feature testing if a pingsource becomes ready.
-func PingSourceGoesReady(name string, cfg ...pingsource.CfgFn) *feature.Feature {
+// GoesReady returns a feature testing if a pingsource becomes ready.
+func GoesReady(name string) *feature.Feature {
 	f := feature.NewFeatureNamed("PingSource goes ready.")
-
-	sink := feature.MakeRandomK8sName("sink")
-	f.Setup("install a service", svc.Install(sink, "app", "rekt"))
-
-	cfg = append(cfg, pingsource.WithSink(&duckv1.KReference{
-		Kind:       "Service",
-		Name:       sink,
-		APIVersion: "v1",
-	}, ""))
-
-	f.Setup("install a pingsource", pingsource.Install(name, cfg...))
 
 	f.Stable("pingsource").
 		Must("be ready", pingsource.IsReady(name, features.Interval, features.Timeout))
